# Integrate ripgrep for optimized search performance