--- a/src/search/mod.rs
+++ b/src/search/mod.rs
@@ -14,6 +14,7 @@ pub mod search_tokens;
 pub mod term_exceptions; // New module for term exceptions
 pub mod timeout; // New module for timeout functionality
 pub mod tokenization; // New elastic search query parser
+pub mod ripgrep_searcher; // New high-performance ripgrep-based searcher
                       // Temporarily commented out due to compilation issues
                       // mod temp_frequency_search;