… CLAUDE.md
… Cargo.toml
… site/blog/v0.6.0-release.md
… site/changelog.md